+++ test/gtest/opeq_test.cpp
@@ -15,7 +15,6 @@ limitations under the License.
 */
 
 #include "gtest/gtest.h"
-
 #include "ir/ir.h"
 #include "ir/visitor.h"
 #include "lib/exceptions.h"
